+++ src/redturtle/prenotazioni/restapi/services/booking/vacation.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+
+### DEPRECATED ####
+
 from plone.protect.interfaces import IDisableCSRFProtection
 from plone.restapi.deserializer import json_body
 from plone.restapi.services import Service